Update render function to use UInt32 for rawCommands
Change rawCommands type from Float to UInt32 for better compatibility with binary command streams.

+++ AlloyCore/Renderer/AlloyRenderer.swift
@@ -28,15 +28,16 @@ public class AlloyRenderer {
         }
     }
     
-    public func render(drawable: CAMetalDrawable, texture: MTLTexture, rawCommands: [Float]) {
+    // 🔥 接收 UInt32 二进制指令流
+    public func render(drawable: CAMetalDrawable, texture: MTLTexture, rawCommands: [UInt32]) {
         let drawableTexture = drawable.texture
         guard !rawCommands.isEmpty else { return }
         
         let commandBuffer = device.makeBuffer(bytes: rawCommands,
-                                              length: rawCommands.count * MemoryLayout<Float>.size,
+                                              length: rawCommands.count * MemoryLayout<UInt32>.size,
                                               options: .storageModeShared)
         
-        // 🔥 步长从 27 变成 36 (6 pos + 12 color + 6 uv + 3 z + 9 normal)
+        // 步长依然是 36 个 32位数据
         var commandCount = UInt32(rawCommands.count / 36)
         
         guard let cmdQueueBuffer = commandQueue.makeCommandBuffer(),
